Move aggregation constants and refactor
