+++ src/constants/metadata.js
@@ -1,0 +1,3 @@
+export const AGG_PREFIX = '_cdb_agg_';
+export const DIM_PREFIX = '_cdb_dim_';
+export const CLUSTER_FEATURE_COUNT = '_cdb_feature_count';
